requirements.txt
streamlit>=1.26.0
torch>=2.0.0
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.23.0
pillow>=10.0.0
joblib>=1.3.0
